Ditched Parallelization & added Configuration-Readme File
Resistance is futile, parallelize TRACS possible is not.
After several weeks of fight we have decided not to parallelize TRACS. We have encoutered errors in both FENICS/DOLFIN own libraries (which we fixed) but also on daughter libraries inside of FENICS such as petsc.

A configuration file has been added. This "Config.TRACS" serves both as a simple way to change the simulation parameters without recompiling the code and also as a brief description of how the program works. Depending on the feedback this file might be split in too (Config.TRACS & HowTRACSWork.TRACS) but it's just a possibility.

We will now focus on writting a minimization program that will allow TRACS to calculate Neff & trapping time given some waveforms. Also GUI improvements might also happen soon.