Empty proof-of-concept GUI including irradiation and full simulation
The GUI has been modified to include irradiation parameters and the ability to run a full simulation i.e.: same as running TRACS in CLI-mode but with custom parameters in stead of the ones in Config.TRACS.

The functionality of the GUI has not been modified and the new items are completely empty and do nothing for the moment. Backend support for the new GUI will come.