Implemented ability to modify Neff via Config.TRACS file
The Neff from which the electric potential and field are calculated in the irradiated case can now be modify from run to run. The Neff parametrization is now set up by SMSDetector before it is used.

The parametrization is done via 8 parameters of which two are fixed to be the end and the begining of the detector in th Z-axis.

For further information about Neff there's a brief explanation in Config.TRACS and there will be a extended description in the README in the future.