New TRACSInterface that lets TRACS behave like an API for external usage
New class "TRACSInterface" that allows TRACS to be externally used and embeded by any software.

Currently it offers
	- Methods for changing the voltages and shifting carriers in Z and Y directions.
	- Methods for changing the trapping time and Neff parametrization
	- Simulate the theoretical current (without output)
	- 3 Different outputs in the form of 1D root histograms. Outputs are the ideal(ramo) current, the RC-shaped current and the signal convoluted with RC+amplifier circuit.

This Interface is the first step towards the minimization code but has been designed with flexibility and portability in mind so that any other software that might want to take advantage of TRACS simulations can do so in a very simple manner.