Added on-demand Neff parametrization
Source class and SMSDetector have been updated to take a new input that defines the type of Neff parametrization to be used. The user can select from Linear, Triconstant or Trilinear (more information of the meaning of this terms will be found soon in Config.TRACS and GUIdelines.TRACS)

The number of parameters given to Source is still 8 in any of the three cases and the class itself will decide which ones to use based on the Neff parametrization used.

Linear uses 2 fixed and 2 free parameters. Only extremal points are used.
Triconstant uses 2 fixed and 5 free parameters. All but the last y3 = Neff(z3) are used.
Trilinear uses 2 fixed and 6 free parameters. Its the same implementation we previously had.