Updated GUI and Improved the Custom Neff Visualization Tool (CNVT)
The GUI now allows for diferent Neff parametrization, catching up with CLI-version update form last commit.

The greatest feature of the TRACS-GUI radiation upgrade, the CNVT, just got even better with up to 3 different parametrization options for the custom Neff definition. The CNVT has also received its own dedicated tab inside the Potentials page and a resizable space that allows the user to enlarge or shrink the graph on demand.
CNVT is the greatest feature of TRACS-GUI upgrade and allows users to modify Neff on the fly while still maintaining full control over the process through the fast feedback loop that a graphical representation of the input parameters allows. Now tha possibilities have widen with up to 3 different parametrization in the same compact and familiar tab, just bigger and better.

Users should expect more parametrizations to be implemented in the near future to allow for maximum flexibility and better accuracy when parametrazin virtually ANY Neff.

Get on the Hype TRAIN!