Included Neff On-the-fly parametrization choser as configuration parameter
The Neff parametrization is now an input parameter in the Config.TRACS file. Explanation has been added as well.
The parser now recognises neffType as an input parameter and is able to read and set it properly.
main.cpp also makes use of nefftype and creates a detector with different values depending on the configuration value.
TRACSInterface has been updated to alse be able to change neffType in detector via a simple setter.
TRACSInterface is able now to change carriers file in the middle of the simulation, should the user desire.
GUIdelines.TRACS has been deleted as its text is already present in README.md