Fixed startup appearance and added carriers shifting to GUI
The appearance of the GUI on startup is fixed so that is visually more pleasent whilst maintaining the sliding fuctionality that allows almost any graph to be full window size.
Carrier collection can now be moved (shifted) around using GUI only. This reduces the number of required laser simulation by a lot as now one file can be moved around to simulate any ilumination position